requirements.txt
fastapi>=0.85.0
Pillow>=9.2.0
PyYAML>=6.0
aiogram>=2.22.2
websockets>=10.3
torch>=1.12.1
torchvision>=0.13.1
pytest>=7.1.3
uvicorn>=0.18.3
